Feature diameter in GAL file is double
git-svn-id: https://subversion.biologie.ens.fr/svn/leburon/trunk/nividic@282 5a62d92b-e016-0410-9103-b89fe532bb34

+++ src/main/java/fr/ens/transcriptome/nividic/om/GenepixArrayList.java
@@ -46,8 +46,10 @@ public class GenepixArrayList {
   private static final String GAL_HEADER_URL = "URL";
 
   private static final String GAL_HEADER_SUPPLIER = "Supplier";
-  private static final String GAL_HEADER_ARRAYERSOFTWARENAME = "ArrayerSoftwareName";
-  private static final String GAL_HEADER_ARRAYERSOFTWAREVERSION = "ArrayerSoftwareVersion";
+  private static final String GAL_HEADER_ARRAYERSOFTWARENAME =
+      "ArrayerSoftwareName";
+  private static final String GAL_HEADER_ARRAYERSOFTWAREVERSION =
+      "ArrayerSoftwareVersion";
   private static final String GAL_HEADER_ARRAYNAME = "ArrayName";
   private static final String GAL_HEADER_ARRAYREVISON = "ArrayRevison";
   private static final String GAL_HEADER_SLIDEBARCODE = "SlideBarCode";
@@ -511,7 +513,7 @@ public class GenepixArrayList {
         return ab;
 
       if (st.hasMoreTokens())
-        ab.setFeatureDiameter(Integer.parseInt(st.nextToken().trim()));
+        ab.setFeatureDiameter(Double.parseDouble(st.nextToken().trim()));
       else
         return ab;
